Clean up moving api ext into its own module
Turns out we can use async_trait for the ApiExt trait
